components: rename entity_mentions to github_integration
This also adds a "mentions" subpackage.
