+++ app/components/__init__.py
@@ -4,7 +4,7 @@ from . import (
     autoclose,
     close_help_post,
     docs,
-    entity_mentions,
+    github_integration,
     message_filter,
     move_message,
     status,
@@ -17,7 +17,7 @@ __all__ = (
     "autoclose",
     "close_help_post",
     "docs",
-    "entity_mentions",
+    "github_integration",
     "message_filter",
     "move_message",
     "status",
