+++ app/components/github_integration/__init__.py
@@ -8,13 +8,14 @@ from .comments import (
     entity_comment_edit_hook,
     reply_with_comments,
 )
-from .fmt import entity_message, load_emojis
-from .integration import (
+from .mentions import (
+    ENTITY_REGEX,
     entity_mention_delete_hook,
     entity_mention_edit_hook,
+    entity_message,
+    load_emojis,
     reply_with_entities,
 )
-from .resolution import ENTITY_REGEX
 
 __all__ = (
     "ENTITY_REGEX",
