+++ app/components/github_integration/mentions/cache.py
@@ -1,7 +1,7 @@
 from githubkit.exception import RequestFailed
 
 from .discussions import get_discussion
-from .models import Entity, Issue, PullRequest
+from app.components.github_integration.models import Entity, Issue, PullRequest
 from app.setup import gh
 from app.utils import TTRCache
